Structure change and refactor
