+++ web_contents/app.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request
-from data_fetcher import fetch_semesters, fetch_subjects
+from web_contents.data.data_fetcher import fetch_semesters, fetch_subjects
 
 app = Flask(__name__)
 app.debug = True
