+++ web_contents/data/data_processor.py
@@ -5,9 +5,10 @@ import numpy as np
 import pandas as pd
 from bs4 import BeautifulSoup
 from typing import List
-from constants import base_url, ClassConx, ClassExam, ClassNumber, \
-    ClassInstructor, SeatInfo, SKIP_BEGINNING
-from data_fetcher import fetch_web_content, fetch_subjects, fetch_semesters
+from web_contents.data.constants import base_url, ClassConx, \
+    ClassExam, ClassNumber, ClassInstructor, SeatInfo, SKIP_BEGINNING
+from web_contents.data.data_fetcher import fetch_semesters, fetch_subjects, \
+    fetch_web_content
 
 
 def extract_class_info(web_content: str) -> List[list]:
@@ -296,10 +297,10 @@ def get_semester_class_info(semester_name: str, semester_value: str):
     semester_frame = pd.DataFrame(pd.concat(all_class, ignore_index=True))
 
     # Save it as a pickle file.
-    semester_frame.to_pickle(f"pickle_data/{semester_name}.pkl")
+    semester_frame.to_pickle(f"saved_data/pickle_data/{semester_name}.pkl")
 
     # Save to CSV in order to easily compare with web page.
-    semester_frame.to_csv(f"csv_data/{semester_name}.csv")
+    semester_frame.to_csv(f"saved_data/csv_data/{semester_name}.csv")
 
 
 def save_all_info():
