+++ web_contents/data/data_reader.py
@@ -2,7 +2,7 @@
 """This file reads the data and select right query based on users request."""
 
 import pandas as pd
-from data_processor import save_all_info
+from web_contents.data.data_processor import save_all_info
 # from constants import ClassConx, ClassInstructor, ClassExam, ClassNumber
 # TODO: See if I want to create a fake function to get rid of stupid warnings
 # TODO: read documentation about queries
